+++ src/action/apply-suggestions.ts
@@ -2,8 +2,8 @@ import { access, mkdir, readFile, writeFile } from 'node:fs/promises';
 import { dirname } from 'node:path';
 import type { Remote } from 'comlink';
 import type { ESLint } from 'eslint';
-import { fixingSpinner } from '../cli/ora.js';
 import { promptToInputReuseFilterScript } from '../cli/prompt.js';
+import { fixingSpinner } from '../cli/spinner.js';
 import type { Undo } from '../core.js';
 import type { SerializableCore } from '../core-worker.js';
 import {
