+++ src/action/fix.ts
@@ -1,6 +1,6 @@
 import type { Remote } from 'comlink';
 import type { ESLint } from 'eslint';
-import { fixingSpinner } from '../cli/ora.js';
+import { fixingSpinner } from '../cli/spinner.js';
 import type { Undo } from '../core.js';
 import type { SerializableCore } from '../core-worker.js';
