+++ src/scene/lint.ts
@@ -1,6 +1,6 @@
 import type { Remote } from 'comlink';
 import { warn } from '../cli/log.js';
-import { lintingSpinner } from '../cli/ora.js';
+import { lintingSpinner } from '../cli/spinner.js';
 import type { SerializableCore } from '../core-worker.js';
 import { unique } from '../util/array.js';
 import { notEmpty } from '../util/type-check.js';
